# Optimise createTable in stream_writer.go